requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-appconf==1.0.5
django-cors-headers==3.8.0
django-dotenv==1.4.2
django-imagekit==4.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
pilkit==2.0
Pillow==8.3.2
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
